// microfrontends/api-middleware/src/Profile/action-types.ts
export enum ProfileActionType {
  API_PROFILE_FETCH_DATA = "API/PROFILE/FETCH_DATA",
  PROFILE_SET_DATA = "PROFILE_SET_DATA",
  PROFILE_FETCH_START = "PROFILE_FETCH_START",
  PROFILE_FETCH_COMPLETE = "PROFILE_FETCH_COMPLETE",

  API_PROFILE_PUT_DATA = "API/PROFILE/PUT_DATA",
  PROFILE_PUT_START = "PROFILE_PUT_START",
  PROFILE_PUT_COMPLETE = "PROFILE_PUT_COMPLETE",
}